fix: 회원가입 플로우 수정
